renamed DetectionPaddingProcessor to DetectionTransformationProcessor
